Moved obs list converters to avoid import problems with datastore.
